requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
Faker==24.9.0
openpyxl==3.1.2
python-dotenv==1.0.0
